requirements.txt
# Core dependencies
python-chess>=1.999
numpy>=1.21.0
typing-extensions>=4.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional dependencies for advanced features
stockfish>=3.28.0  # For testing against strong engine
matplotlib>=3.5.0  # For visualization
pandas>=1.3.0      # For game analysis